[Refactor] HpMovingBox의 일부 코드를 SelectedContent로 옮김
